+++ src/Symfony/Component/Serializer/Normalizer/UnwrappingDenormalizer.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Serializer\Normalizer;
 
 use Symfony\Component\PropertyAccess\PropertyAccess;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
+use Symfony\Component\Serializer\DenormalizationResult;
 use Symfony\Component\Serializer\SerializerAwareInterface;
 use Symfony\Component\Serializer\SerializerAwareTrait;
 
@@ -48,7 +49,15 @@ final class UnwrappingDenormalizer implements DenormalizerInterface, SerializerA
             $data = $this->propertyAccessor->getValue($data, $propertyPath);
         }
 
-        return $this->serializer->denormalize($data, $class, $format, $context);
+        $result = $this->serializer->denormalize($data, $class, $format, $context);
+
+        if ($result instanceof DenormalizationResult && !$result->isSucessful()) {
+            $propertyPath = str_replace('][', '.', substr($propertyPath, 1, -1));
+
+            return DenormalizationResult::failure($result->getInvariantViolationsNestedIn($propertyPath));
+        }
+
+        return $result;
     }
 
     /**