+++ src/Symfony/Component/Serializer/SerializerInterface.php
@@ -30,6 +30,10 @@ interface SerializerInterface
     /**
      * Deserializes data into the given type.
      *
+     * When context option `collect_invariant_violations` is enabled, the
+     * serializer must always return an instance of
+     * {@see \Symfony\Component\Serializer\DenormalizationResult}.
+     *
      * @param mixed $data
      *
      * @return mixed